Revert "zmiannn7yy"
