+++ Hello.go
@@ -1,0 +1,16 @@
+package main
+
+//tersdsa
+import "fmt"
+
+func main() {
+	var par string
+	fmt.Println("Jaka parówa wariacie? K/S")
+	fmt.Scanln(&par)
+	if par == "K" {
+		fmt.Println("Kabanos ♥")
+	} else {
+		fmt.Println("ser :CCCC")
+	}
+
+}
