+++ go.mod
@@ -1,0 +1,3 @@
+module github.com/Suchydudul/GolangTest
+
+go 1.16
