--- new/go.mod
@@ -1,3 +1,0 @@
-module github.com/Suchydudul/GolangTest/new
-
-go 1.16
